cub3d 01 04 2021

+++ cub3d_srcs/functions/check_raycol_direction.c
@@ -6,7 +6,7 @@
 /*   By: sachouam <sachouam@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/25 02:12:53 by sachouam          #+#    #+#             */
-/*   Updated: 2021/03/19 17:38:18 by sachouam         ###   ########.fr       */
+/*   Updated: 2021/04/01 11:04:20 by sachouam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,21 +15,25 @@
 int
 	ft_check_raycol_direction(t_all *all)
 {
-	if (all->vect.raycol < PI / 2 && all->vect.raycol > 0.0)
+	if ((float)all->vect.raycol < (float)(PI / 2)
+	&& (float)all->vect.raycol > 0.0)
 		return (1);
-	else if (all->vect.raycol > PI / 2 && all->vect.raycol < PI)
+	else if ((float)all->vect.raycol > (float)(PI / 2)
+	&& (float)all->vect.raycol < (float)(PI))
 		return (2);
-	else if (all->vect.raycol > PI && all->vect.raycol < PI + (PI / 2))
+	else if ((float)all->vect.raycol > (float)(PI)
+	&& (float)all->vect.raycol < (float)(PI + (PI / 2)))
 		return (3);
-	else if (all->vect.raycol > PI + (PI / 2) && all->vect.raycol < 2 * PI)
+	else if ((float)all->vect.raycol > (float)(PI + (PI / 2))
+	&& (float)all->vect.raycol < (float)(2 * PI))
 		return (4);
-	else if (all->vect.raycol == PI * 2)
+	else if ((float)all->vect.raycol == (float)(PI * 2))
 		return (5);
-	else if (all->vect.raycol == PI / 2)
+	else if ((float)all->vect.raycol == (float)(PI / 2))
 		return (6);
-	else if (all->vect.raycol == PI)
+	else if ((float)all->vect.raycol == (float)(PI))
 		return (7);
-	else if (all->vect.raycol == PI + (PI / 2))
+	else if ((float)all->vect.raycol == (float)(PI + (PI / 2)))
 		return (8);
 	return (0);
 }
